changed time fields

+++ src/main/kotlin/io/github/x1111101101/glucoseserver/record/model/TokDailyRecord.kt
@@ -4,7 +4,7 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class TokDailyRecord (
-    val date: Long
+    val date: Int
 ) {
 
 }
